# docs/swagger.yaml
basePath: /v1
definitions:
...
  swagger.DatesResponse:
    properties:
      dates:
        example:
        - "2024-02-25"
        - "2024-02-26"
        - "2024-03-25"
        items:
          type: string
        type: array
    type: object
...
  swagger.ForbiddenError:
    properties:
      error:
        example: forbidden
        type: string
    type: object
  swagger.ForbiddenResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.ForbiddenError'
        type: array
    type: object
  swagger.GarminDaily:
    properties:
      activeKilocalories:
        example: 0
        type: integer
      activeTimeInSeconds:
        example: 0
        type: integer
      activityType:
        example: GENERIC
        type: string
      bmrKilocalories:
        example: 1967
        type: integer
      calendarDate:
        example: "2025-02-20"
        type: string
      distanceInMeters:
        example: 0
        type: number
      durationInSeconds:
        example: 76020
        type: integer
      floorsClimbedGoal:
        example: 10
        type: integer
      intensityDurationGoalInSeconds:
        example: 36000
        type: integer
      moderateIntensityDurationInSeconds:
        example: 0
        type: integer
      startTimeInSeconds:
        example: 1740002400
        type: integer
      startTimeOffsetInSeconds:
        example: 7200
        type: integer
      steps:
        example: 0
        type: integer
      stepsGoal:
        example: 10000
        type: integer
      summaryId:
        example: x32c7d85-67b65460-128f4-0
        type: string
      vigorousIntensityDurationInSeconds:
        example: 0
        type: integer
    type: object
  swagger.GarminDailyExample:
    properties:
      calendarDate:
        example: "2025-06-12"
        type: string
      steps:
        example: 5000
        type: integer
    type: object
  swagger.GarminData:
    properties:
      dailies:
        $ref: '#/definitions/swagger.GarminDaily'
      epochs:
        $ref: '#/definitions/swagger.GarminEpoch'
    type: object
  swagger.GarminDataResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.GarminData'
    type: object
  swagger.GarminEpoch:
    properties:
      activeKilocalories:
        example: 0
        type: integer
      activeTimeInSeconds:
        example: 0
        type: integer
      activityType:
        example: UNMONITORED
        type: string
      distanceInMeters:
        example: 0
        type: number
      durationInSeconds:
        example: 900
        type: integer
      intensity:
        example: SEDENTARY
        type: string
      maxMotionIntensity:
        example: 0
        type: number
      meanMotionIntensity:
        example: 0
        type: number
      met:
        example: 1
        type: number
      startTimeInSeconds:
        example: 1740009600
        type: integer
      startTimeOffsetInSeconds:
        example: 7200
        type: integer
      steps:
        example: 0
        type: integer
      summaryId:
        example: x32c7d85-67b67080-9
        type: string
    type: object
  swagger.GarminPayloadExample:
    properties:
      dailies:
        items:
          $ref: '#/definitions/swagger.GarminDailyExample'
        type: array
    type: object
  swagger.GarminPostDataInput:
    properties:
      data:
        $ref: '#/definitions/swagger.GarminPayloadExample'
      date:
        example: "2025-06-12"
        type: string
      user_id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
    type: object
...
  swagger.GarminTypesResponse:
    properties:
      types:
        example:
        - epochs
        - ' dailies'
        - ' hrv'
        items:
          type: string
        type: array
    type: object
...
  swagger.InternalServerError:
    properties:
      error:
        example: the server encountered a problem
        type: string
    type: object
  swagger.InternalServerErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.InternalServerError'
        type: array
    type: object
  swagger.InvalidDateRange:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.UnprocessableEntityError'
        type: array
    type: object
...
  swagger.ServiceUnavailableError:
    properties:
      error:
        example: database is unavailable
        type: string
    type: object
  swagger.ServiceUnavailableResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.ServiceUnavailableError'
        type: array
    type: object
...
  swagger.UnauthorizedError:
    properties:
      error:
        example: unauthorized
        type: string
    type: object
  swagger.UnauthorizedResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.UnauthorizedError'
        type: array
    type: object
  swagger.UnprocessableEntityError:
    properties:
      error:
        example: invalid date range
        type: string
    type: object
...
  swagger.ValidationError:
    properties:
      parameter name:
        example: full description of the error
        type: string
    type: object
  swagger.ValidationErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.ValidationError'
        type: array
    type: object
info:
  contact: {}
  description: API for integrating, analyzing, and visualizing sports and health data
  termsOfService: https://csc.fi/en/security-privacy-data-policy-and-open-source-policy/privacy/
  title: KUHA REST API
paths:
...
  /utv/garmin/data:
    delete:
      consumes:
      - application/json
      description: Deletes all Garmin data entries for a specific user
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Deleted: Data successfully removed'
        "204":
          description: 'No Content: No matching data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Delete all Garmin data for a user
      tags:
      - UTV - Garmin
    get:
      consumes:
      - application/json
      description: Returns data for the specified user on the specified date (optionally
        filtered by key)
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      - description: Date (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      - description: Type
        in: query
        name: key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data
          schema:
            $ref: '#/definitions/swagger.GarminDataResponse'
        "204":
          description: 'No Content: No data found'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Get available data
      tags:
      - UTV - Garmin
    post:
      consumes:
      - application/json
      description: Posts Garmin data for the specified user on the specified date
      parameters:
      - description: Garmin data input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/swagger.GarminPostDataInput'
      produces:
      - application/json
      responses:
        "201":
          description: 'Created: Data successfully stored (no content in response
            body)'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Post Garmin data
      tags:
      - UTV - Garmin
  /utv/garmin/dates:
    get:
      consumes:
      - application/json
      description: Returns available dates for the specified user (optionally filtered
        by date range)
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      - description: Filter dates after this date (YYYY-MM-DD)
        in: query
        name: after_date
        type: string
      - description: Filter dates before this date (YYYY-MM-DD)
        in: query
        name: before_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of available dates
          schema:
            $ref: '#/definitions/swagger.DatesResponse'
        "204":
          description: 'No Content: No available dates found'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.InvalidDateRange'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Get available dates
      tags:
      - UTV - Garmin
...
  /utv/garmin/types:
    get:
      consumes:
      - application/json
      description: Returns available types for the specified user on the specified
        date
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      - description: Date (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of available types
          schema:
            $ref: '#/definitions/swagger.GarminTypesResponse'
        "204":
          description: 'No Content: No available types found'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Get available types
      tags:
      - UTV - Garmin
...
securityDefinitions:
  BearerAuth:
    description: 'Use format: Bearer your_JWT_here'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"